fix: 🐛 remove json and v2 files from components and static
