+++ src/components/MoCATable.jsx
@@ -5,32 +5,14 @@ import FilterableTable from 'react-filterable-table';
 // eslint-disable-next-line import/no-unresolved
 import BrowserOnly from '@docusaurus/BrowserOnly';
 
-import MappingsJSON from '../../../docs/static/json/mappings.json';
+import MoCA from '../../docs/static/json/moCA.json';
 
-export default function OmopMappingTable() {
-  const targetConceptIdRender = (props) => {
-    const { value, record } = props;
-
-    const isLocalConceptId = record.TARGET_CONCEPT_ID === record.LOCAL_CONCEPT_ID;
-
-    return isLocalConceptId ? (
-      <span>{value}</span>
-    ) : (
-      <a
-        href={`https://athena.ohdsi.org/search-terms/terms/${props.value}`}
-        target="_blank"
-        rel="noreferrer"
-      >
-        {props.value}
-      </a>
-    );
-  };
-
+export default function StructureTable() {
   // Fields to show in the table, and what object properties in the data they bind to
   const fields = [
     {
-      name: 'Protected or open-source?',
-      displayName: 'Protected or open-source?',
+      name: 'Protected_or_open-source',
+      displayName: 'Protected_or_open-source',
       inputFilterable: true,
     },
     {
@@ -39,18 +21,13 @@ export default function OmopMappingTable() {
       inputFilterable: true,
     },
     {
-      name: 'FIELD_TYPE',
-      displayName: 'FIELD_TYPE',
+      name: 'MOCA_FILE_FIELDNAME',
+      displayName: 'MOCA_FILE_FIELDNAME',
       inputFilterable: true,
     },
     {
-      name: 'FIELD_ID',
-      displayName: 'FIELD_ID',
-      inputFilterable: true,
-    },
-    {
-      name: 'FIELD_ID_NUM',
-      displayName: 'FIELD_ ID_NUM',
+      name: 'MOCA_FILE_FIELDNAME_NAMING_UNIT',
+      displayName: 'MOCA_FILE_FIELDNAME_NAMING_UNIT',
       inputFilterable: true,
     },
     {
@@ -59,79 +36,88 @@ export default function OmopMappingTable() {
       inputFilterable: true,
     },
     {
-      name: 'SRC_CODE_ID',
-      displayName: 'SRC_CODE_ID',
+      name: 'SRC_CODE_DESCRIPTION',
+      displayName: 'SRC_CODE_DESCRIPTION',
       inputFilterable: true,
     },
     {
-      name: 'SRC_CD_DESCRIPTION',
-      displayName: 'SRC_CD_ DESCRIPTION',
+      name: 'UCUM_CONCEPT_ID',
+      displayName: 'UCUM_CONCEPT_ID',
       inputFilterable: true,
     },
     {
-      name: 'Choices, Calculations, OR Slider Labels From REDCap CODEBOOK',
-      displayName: 'Choices, Calculations, OR Slider Labels From REDCap CODEBOOK',
+      name: 'Example_Data',
+      displayName: 'Example_Data',
+      inputFilterable: true,
+    },
+    {
+      name: 'Data_Type',
+      displayName: 'Data_Type',
+      inputFilterable: true,
+    },
+    {
+      name: 'Value_Range',
+      displayName: 'Value_Range',
+      inputFilterable: true,
+    },
+    {
+      name: 'Data_Description',
+      displayName: 'Data_Description',
+      inputFilterable: true,
+    },
+    {
+      name: 'Equivalent_NDA_MOCA_element_name',
+      displayName: 'Equivalent_NDA_MOCA_element_name',
       inputFilterable: true,
     },
     {
       name: 'TARGET_CONCEPT_ID',
-      displayName: 'TARGET_ CONCEPT_ID',
+      displayName: 'TARGET_CONCEPT_ID',
       inputFilterable: true,
-      render: targetConceptIdRender,
     },
     {
       name: 'LOCAL_CONCEPT_ID',
-      displayName: 'LOCAL _CONCEPT_ID',
+      displayName: 'LOCAL_CONCEPT_ID',
       inputFilterable: true,
     },
     {
       name: 'TARGET_CONCEPT_NAME',
-      displayName: 'TARGET_ CONCEPT_NAME',
+      displayName: 'TARGET_CONCEPT_NAME',
       inputFilterable: true,
     },
     {
       name: 'TARGET_DOMAIN_ID',
-      displayName: 'TARGET_ DOMAIN_ID',
+      displayName: 'TARGET_DOMAIN_ID',
       inputFilterable: true,
     },
     {
       name: 'TARGET_VOCABULARY_ID',
-      displayName: 'TARGET_ VOCABULARY_ ID',
+      displayName: 'TARGET_VOCABULARY_ID',
       inputFilterable: true,
     },
     {
       name: 'TARGET_CONCEPT_CLASS_ID',
-      displayName: 'TARGET_ CONCEPT _CLASS_ID',
+      displayName: 'TARGET_CONCEPT_CLASS_ID',
       inputFilterable: true,
     },
     {
       name: 'TARGET_STANDARD_CONCEPT',
-      displayName: 'TARGET_ STANDARD _CONCEPT',
+      displayName: 'TARGET_STANDARD_CONCEPT',
       inputFilterable: true,
     },
     {
       name: 'TARGET_CONCEPT_CODE',
-      displayName: 'TARGET_ CONCEPT_CODE',
+      displayName: 'TARGET_CONCEPT_CODE',
       inputFilterable: true,
     },
     {
       name: 'PREDICATE_ID',
-      displayName: 'PREDICATE _ID',
+      displayName: 'PREDICATE_ID',
       inputFilterable: true,
     },
     {
       name: 'CONFIDENCE',
       displayName: 'CONFIDENCE',
-      inputFilterable: true,
-    },
-    {
-      name: 'qualifier_concept_id',
-      displayName: 'qualifier_concept_id',
-      inputFilterable: true,
-    },
-    {
-      name: 'qualifier_source_value',
-      displayName: 'qualifier_source_value',
       inputFilterable: true,
     },
   ];
@@ -140,15 +126,15 @@ export default function OmopMappingTable() {
     <BrowserOnly>
       {() => (
         <div className="filtered-table">
-          <h1 className="header">OMOP Mapping Table</h1>
+          <h1 className="header">MOCA Data Dictionary Table</h1>
           <FilterableTable
             className="filteredTable"
             initialSort="Data Element"
-            data={MappingsJSON}
+            data={MoCA}
             fields={fields}
             noRecordsMessage="No match found."
             noFilteredRecordsMessage="No match found."
-            pageSizes={[10, 30, 50, 100, 1667]}
+            pageSizes={[10, 20, 36]}
             pageSize={10}
           />
         </div>
